Generic Graph Drynamic Input Support
Generic Graph Controller now properly accepts dynamic input.
TODO: Need to triple check data it is using but it appears production is backwards on the data it uses and displays between on/off peak

Code formatting changes from 2 to 4 spaces for a tab.
Documentation changes
Added TODO's into the Controller and Form files.